4-cassandra/py-cassandra-demo/requirements.txt
attrs==26.1.0
Automat==25.4.16
cassandra-driver==3.29.3
click==8.3.2
colorama==0.4.6
constantly==23.10.4
geomet==1.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
packaging==26.0
Twisted==25.5.0
typing_extensions==4.15.0
zope.interface==8.2
